Retire the option module's preamble
Every child is rewritten, so the parent's transpiled preamble is no
longer everybody's import list and what is left in it can finally be
shown dead. It holds no code now — only the vocabulary more than one
child needs — and the module docs carry the child table.

The seam-finding rule from the earlier prunes held again: prune, carve,
rewrite, prune again. The first prune could only reach what no child
named; this one reached the rest.

- The `unsafe extern "C" { }` forward declarations of `vim_regexec`,
  `on_scrollback_option_changed` and `ll_resize_stack` are gone. All
  three are Rust functions in this tree; the children import them.
- The `OPT_*` scope bits are `c_int` rather than `c_uint`, which is the
  type every `opt_flags` argument has. That retired ~60 `as c_int` casts
  and the six per-file `const OPT_X: c_int = super::OPT_X as c_int;`
  shims the rewrites had grown, and the set now carries a doc comment
  saying what each bit means — including that neither `OPT_LOCAL` nor
  `OPT_GLOBAL` means "both", which is what a bare `:set` does.
- Six message texts moved to the child that reports them, as `&CStr`
  consts: E518/E520/E992/E521 to `set_cmd.rs`, E1542/E1543 to
  `validate.rs`.
- The ten `p_*_nobin`/`p_*_nopaste` stashes moved to the option that
  owns them — 'binary' in `check.rs`, 'paste' in `paste.rs` — and
  `copy.rs`, which seeds a new buffer from them, imports them from
  there.
- `CPO_BUFOPT`/`CPO_BUFOPTGLOB` to `copy.rs`, `MAX_SEARCH_COUNT` and
  `B_IMODE_LAST` to `validate.rs`, `INC`/`GAP` to `show.rs`.
- `MAXPATHL` is the literal rather than an alias of a `DEFAULT_MAXPATHL`
  with no other user.

Both `option.rs` and `optionstr.rs` carry `#![deny(unsafe_op_in_unsafe_fn)]`
now. Neither can carry `forbid(unsafe_code)`: the attribute binds
descendants, and the descendants are where the raw pointers are.

Ratchet: a pure shrink, `just refresh`.

Oracles: all ten probes byte-identical. `just oldtest test_options
test_options_all test_cpoptions test_set test_startup` — 540 tests, 0
failures. `just apigen --check` green. The module was exercised by hand
after the prune (`:set fillchars=`, `:set listchars=`, `:set binary`,
`:set paste` and back) — a prune's failure mode is silent by
construction, and the metrics would not have caught it.

+++ crates/nvim/src/nvim/option/info.rs
@@ -113,8 +113,7 @@ pub unsafe fn get_all_vimoptions(arena: *mut Arena) -> Dict {
         let mut retval = arena_dict(arena, kOptCount as size_t);
         for opt_idx in kOptAleph..kOptCount {
             let opt = (options.ptr() as *mut vimoption_T).offset(opt_idx as isize);
-            let opt_dict =
-                vimoption2dict(opt, OPT_GLOBAL as c_int, curbuf.get(), curwin.get(), arena);
+            let opt_dict = vimoption2dict(opt, OPT_GLOBAL, curbuf.get(), curwin.get(), arena);
             *retval.items.add(retval.size) = key_value_pair {
                 key: cstr_as_string((*opt).fullname),
                 value: object {
@@ -146,7 +145,7 @@ unsafe fn last_set(
 ) -> sctx_T {
     // SAFETY: the caller's pointers are live for the scopes reached below.
     unsafe {
-        if opt_flags == OPT_GLOBAL as c_int {
+        if opt_flags == OPT_GLOBAL {
             return (*opt).script_ctx;
         }
         let mut script_ctx = sctx_T {
@@ -162,7 +161,7 @@ unsafe fn last_set(
             script_ctx =
                 (*win).w_onebuf_opt.wo_script_ctx[(*opt).scope_idx[kOptScopeWin as usize] as usize];
         }
-        if opt_flags != OPT_LOCAL as c_int && script_ctx.sc_sid == 0 {
+        if opt_flags != OPT_LOCAL && script_ctx.sc_sid == 0 {
             script_ctx = (*opt).script_ctx;
         }
         script_ctx
